requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
boto3==1.37.4
botocore==1.37.4
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
deep-translator==1.11.4
Django==5.1.7
django-channels==0.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-storages==1.14.5
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environs==11.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
frozenlist==1.5.0
g4f==0.5.0.1
googletrans==3.1.0a0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langdetect==1.0.9
ll==1.0
loguru==0.7.3
magic-filter==1.0.12
marshmallow==3.23.1
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
paypalrestsdk==1.13.3
pillow==11.0.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rpds-py==0.21.0
s3transfer==0.11.3
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
starlette==0.46.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
wheel==0.45.1
yarl==1.18.3
zope.interface==7.2
